Rename Django project name from mysite to hatohol
