+++ client/hatohol/django_realize.py
