+++ client/hatohol/wsgi.py
@@ -16,7 +16,7 @@
 # along with Hatohol. If not, see <http://www.gnu.org/licenses/>.
 
 """
-WSGI config for mysite project.
+WSGI config for hatohol project.
 
 This module contains the WSGI application used by Django's development server
 and any production WSGI deployments. It should expose a module-level variable
@@ -32,7 +32,7 @@ framework.
 """
 import os
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "mysite.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "hatohol.settings")
 
 # This application object is used by any WSGI server configured to use this
 # file. This includes Django's development server, if the WSGI_APPLICATION
